+++ system_modes/changelog.rst
@@ -2,6 +2,12 @@
 Changelog for package system_modes
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.4.0 (2020-09-30)
+-----------
+* mode event now including start and goal mode
+* publish inferred state and mode transitions
+* https://github.com/micro-ROS/system_modes/issues/42
+
 0.3.0 (2020-07-23)
 -----------
 * removed boost dependencies (was: program options)